refactor: standardize ID types to use Long in models - Change AuditionPost userId and id from int to Long - Change Comment postId and id from int to Long - Update test helper methods to use long parameters - Update test assertions for null default values

+++ src/main/java/com/audition/model/AuditionPost.java
@@ -13,10 +13,10 @@ import lombok.extern.jackson.Jacksonized;
 public class AuditionPost {
 
     @Schema(description = "ID of the user who created the post", example = "1")
-    int userId;
+    Long userId;
 
     @Schema(description = "Unique identifier of the post", example = "1")
-    int id;
+    Long id;
 
     @Schema(description = "Title of the post", example = "sunt aut facere repellat provident")
     String title;
