+++ src/test/java/com/audition/service/AuditionServiceTest.java
@@ -168,8 +168,8 @@ class AuditionServiceTest {
         void shouldCallGetPostWithCommentsWhenIncludeCommentsTrue() {
             // Arrange
             final AuditionPost expectedPost = AuditionPost.builder()
-                .id(1)
-                .userId(1)
+                .id(1L)
+                .userId(1L)
                 .title("Test Post 1")
                 .body("Test Body 1")
                 .comments(List.of(createComment(1), createComment(2)))
@@ -261,19 +261,19 @@ class AuditionServiceTest {
         }
     }
 
-    private AuditionPost createPost(final int id) {
+    private AuditionPost createPost(final long id) {
         return AuditionPost.builder()
             .id(id)
-            .userId(1)
+            .userId(1L)
             .title("Test Post " + id)
             .body("Test Body " + id)
             .build();
     }
 
-    private Comment createComment(final int id) {
+    private Comment createComment(final long id) {
         return Comment.builder()
             .id(id)
-            .postId(1)
+            .postId(1L)
             .name("Test Comment " + id)
             .email("test@example.com")
             .body("Comment body " + id)
